+++ jni/jni/libtess2/Source/tess.c
@@ -1,5 +1,5 @@
 /*
-** SGI FREE SOFTWARE LICENSE B (Version 2.0, Sept. 18, 2008)
+** SGI FREE SOFTWARE LICENSE B (Version 2.0, Sept. 18, 2008) 
 ** Copyright (C) [dates of first publication] Silicon Graphics, Inc.
 ** All Rights Reserved.
 **
@@ -9,10 +9,10 @@
 ** to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies
 ** of the Software, and to permit persons to whom the Software is furnished to do so,
 ** subject to the following conditions:
-**
+** 
 ** The above copyright notice including the dates of first publication and either this
 ** permission notice or a reference to http://oss.sgi.com/projects/FreeB/ shall be
-** included in all copies or substantial portions of the Software.
+** included in all copies or substantial portions of the Software. 
 **
 ** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
 ** INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
@@ -20,7 +20,7 @@
 ** BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
 ** TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE
 ** OR OTHER DEALINGS IN THE SOFTWARE.
-**
+** 
 ** Except as contained in this notice, the name of Silicon Graphics, Inc. shall not
 ** be used in advertising or otherwise to promote the sale, use or other dealings in
 ** this Software without prior written authorization from Silicon Graphics, Inc.
@@ -67,15 +67,6 @@ static int LongAxis( TESSreal v[3] )
 
 	if( ABS(v[1]) > ABS(v[0]) ) { i = 1; }
 	if( ABS(v[2]) > ABS(v[i]) ) { i = 2; }
-	return i;
-}
-
-static int ShortAxis( TESSreal v[3] )
-{
-	int i = 0;
-
-	if( ABS(v[1]) < ABS(v[0]) ) { i = 1; }
-	if( ABS(v[2]) < ABS(v[i]) ) { i = 2; }
 	return i;
 }
 
@@ -145,7 +136,7 @@ static void ComputeNormal( TESStesselator *tess, TESSreal norm[3] )
 	if( maxLen2 <= 0 ) {
 		/* All points lie on a single line -- any decent normal will do */
 		norm[0] = norm[1] = norm[2] = 0;
-		norm[ShortAxis(d1)] = 1;
+		norm[LongAxis(d1)] = 1;
 	}
 }
 
@@ -186,7 +177,7 @@ extern int RandomSweep;
 #define S_UNIT_X	(RandomSweep ? (2*drand48()-1) : 1.0)
 #define S_UNIT_Y	(RandomSweep ? (2*drand48()-1) : 0.0)
 #else
-#if defined(SLANTED_SWEEP)
+#if defined(SLANTED_SWEEP) 
 /* The "feature merging" is not intended to be complete.  There are
 * special cases where edges are nearly parallel to the sweep line
 * which are not implemented.  The algorithm should still behave
@@ -295,7 +286,7 @@ void tessProjectPolygon( TESStesselator *tess )
 * (what else would it do??)  The region must consist of a single
 * loop of half-edges (see mesh.h) oriented CCW.  "Monotone" in this
 * case means that any vertical line intersects the interior of the
-* region in a single interval.
+* region in a single interval.  
 *
 * Tessellation consists of adding interior edges (actually pairs of
 * half-edges), to split the region into non-overlapping triangles.
@@ -374,6 +365,7 @@ int tessMeshTessellateMonoRegion( TESSmesh *mesh, TESSface *face )
 	return 1;
 }
 
+
 /* tessMeshTessellateInterior( mesh ) tessellates each region of
 * the mesh which is marked "inside" the polygon.  Each such region
 * must be monotone.
@@ -390,122 +382,8 @@ int tessMeshTessellateInterior( TESSmesh *mesh )
 			if ( !tessMeshTessellateMonoRegion( mesh, f ) ) return 0;
 		}
 	}
+
 	return 1;
-}
-
-
-typedef struct EdgeStackNode EdgeStackNode;
-typedef struct EdgeStack EdgeStack;
-
-struct EdgeStackNode {
-	TESShalfEdge *edge;
-	EdgeStackNode *next;
-};
-
-struct EdgeStack {
-	EdgeStackNode *top;
-	struct BucketAlloc *nodeBucket;
-};
-
-int stackInit( EdgeStack *stack, TESSalloc *alloc )
-{
-	stack->top = NULL;
-	stack->nodeBucket = createBucketAlloc( alloc, "CDT nodes", sizeof(EdgeStackNode), 512 );
-	return stack->nodeBucket != NULL;
-}
-
-void stackDelete( EdgeStack *stack )
-{
-    deleteBucketAlloc( stack->nodeBucket );
-}
-
-int stackEmpty( EdgeStack *stack )
-{
-	return stack->top == NULL;
-}
-
-void stackPush( EdgeStack *stack, TESShalfEdge *e )
-{
-	EdgeStackNode *node = (EdgeStackNode *)bucketAlloc( stack->nodeBucket );
-	if ( ! node ) return;
-	node->edge = e;
-	node->next = stack->top;
-	stack->top = node;
-}
-
-TESShalfEdge *stackPop( EdgeStack *stack )
-{
-	TESShalfEdge *e = NULL;
-	EdgeStackNode *node = stack->top;
-	if (node) {
-		stack->top = node->next;
-		e = node->edge;
-		bucketFree( stack->nodeBucket, node );
-	}
-	return e;
-}
-
-
-//	Starting with a valid triangulation, uses the Edge Flip algorithm to
-//	refine the triangulation into a Constrained Delaunay Triangulation.
-void tessMeshRefineDelaunay( TESSmesh *mesh, TESSalloc *alloc )
-{
-	// At this point, we have a valid, but not optimal, triangulation.
-	// We refine the triangulation using the Edge Flip algorithm
-	//
-	//  1) Find all internal edges
-	//	2) Mark all dual edges
-	//	3) insert all dual edges into a queue
-
-	TESSface *f;
-	EdgeStack stack;
-	TESShalfEdge *e;
-	int maxFaces = 0, maxIter = 0, iter = 0;
-
-	stackInit(&stack, alloc);
-
-	for( f = mesh->fHead.next; f != &mesh->fHead; f = f->next ) {
-		if ( f->inside) {
-			e = f->anEdge;
-			do {
-				e->mark = EdgeIsInternal(e); // Mark internal edges
-				if (e->mark && !e->Sym->mark) stackPush(&stack, e); // Insert into queue
-				e = e->Lnext;
-			} while (e != f->anEdge);
-			maxFaces++;
-		}
-	}
-
-	// The algorithm should converge on O(n^2), since the predicate is not robust,
-	// we'll save guard against infinite loop.
-	maxIter = maxFaces * maxFaces;
-
-	// Pop stack until we find a reversed edge
-	// Flip the reversed edge, and insert any of the four opposite edges
-	// which are internal and not already in the stack (!marked)
-	while (!stackEmpty(&stack) && iter < maxIter) {
-		e = stackPop(&stack);
-		e->mark = e->Sym->mark = 0;
-		if (!tesedgeIsLocallyDelaunay(e)) {
-			TESShalfEdge *edges[4];
-			int i;
-			tessMeshFlipEdge(mesh, e);
-			// for each opposite edge
-			edges[0] = e->Lnext;
-			edges[1] = e->Lprev;
-			edges[2] = e->Sym->Lnext;
-			edges[3] = e->Sym->Lprev;
-			for (i = 0; i < 4; i++) {
-				if (!edges[i]->mark && EdgeIsInternal(edges[i])) {
-					edges[i]->mark = edges[i]->Sym->mark = 1;
-					stackPush(&stack, edges[i]);
-				}
-			}
-		}
-		iter++;
-	}
-
-	stackDelete(&stack);
 }
 
 
@@ -598,7 +476,7 @@ TESStesselator* tessNewTess( TESSalloc* alloc )
 
 	if (alloc == NULL)
 		alloc = &defaulAlloc;
-
+	
 	/* Only initialize fields which can be changed by the api.  Other fields
 	* are initialized where they are used.
 	*/
@@ -619,7 +497,7 @@ TESStesselator* tessNewTess( TESSalloc* alloc )
 		tess->alloc.dictNodeBucketSize = 512;
 	if (tess->alloc.regionBucketSize == 0)
 		tess->alloc.regionBucketSize = 256;
-
+	
 	tess->normal[0] = 0;
 	tess->normal[1] = 0;
 	tess->normal[2] = 0;
@@ -629,10 +507,7 @@ TESStesselator* tessNewTess( TESSalloc* alloc )
 	tess->bmax[0] = 0;
 	tess->bmax[1] = 0;
 
-	tess->reverseContours = 0;
-    
 	tess->windingRule = TESS_WINDING_ODD;
-	tess->processCDT = 0;
 
 	if (tess->alloc.regionBucketSize < 16)
 		tess->alloc.regionBucketSize = 16;
@@ -646,7 +521,7 @@ TESStesselator* tessNewTess( TESSalloc* alloc )
 
 	tess->outOfMemory = 0;
 	tess->vertexIndexCounter = 0;
-
+	
 	tess->vertices = 0;
 	tess->vertexIndices = 0;
 	tess->vertexCount = 0;
@@ -658,9 +533,9 @@ TESStesselator* tessNewTess( TESSalloc* alloc )
 
 void tessDeleteTess( TESStesselator *tess )
 {
-
+	
 	struct TESSalloc alloc = tess->alloc;
-
+	
 	deleteBucketAlloc( tess->regionPool );
 
 	if( tess->mesh != NULL ) {
@@ -739,7 +614,7 @@ void OutputPolymesh( TESStesselator *tess, TESSmesh *mesh, int elementType, int 
 			edge = edge->Lnext;
 		}
 		while (edge != f->anEdge);
-
+		
 		assert( faceVerts <= polySize );
 
 		f->n = maxFaceCount;
@@ -756,7 +631,7 @@ void OutputPolymesh( TESStesselator *tess, TESSmesh *mesh, int elementType, int 
 		tess->outOfMemory = 1;
 		return;
 	}
-
+	
 	tess->vertexCount = maxVertexCount;
 	tess->vertices = (TESSreal*)tess->alloc.memalloc( tess->alloc.userData,
 													 sizeof(TESSreal) * tess->vertexCount * vertexSize );
@@ -773,7 +648,7 @@ void OutputPolymesh( TESStesselator *tess, TESSmesh *mesh, int elementType, int 
 		tess->outOfMemory = 1;
 		return;
 	}
-
+	
 	// Output vertices.
 	for ( v = mesh->vHead.next; v != &mesh->vHead; v = v->next )
 	{
@@ -795,7 +670,7 @@ void OutputPolymesh( TESStesselator *tess, TESSmesh *mesh, int elementType, int 
 	for ( f = mesh->fHead.next; f != &mesh->fHead; f = f->next )
 	{
 		if ( !f->inside ) continue;
-
+		
 		// Store polygon
 		edge = f->anEdge;
 		faceVerts = 0;
@@ -864,7 +739,7 @@ void OutputContours( TESStesselator *tess, TESSmesh *mesh, int vertexSize )
 		tess->outOfMemory = 1;
 		return;
 	}
-
+	
 	tess->vertices = (TESSreal*)tess->alloc.memalloc( tess->alloc.userData,
 													  sizeof(TESSreal) * tess->vertexCount * vertexSize );
 	if (!tess->vertices)
@@ -880,7 +755,7 @@ void OutputContours( TESStesselator *tess, TESSmesh *mesh, int vertexSize )
 		tess->outOfMemory = 1;
 		return;
 	}
-
+	
 	verts = tess->vertices;
 	elements = tess->elements;
 	vertInds = tess->vertexIndices;
@@ -976,24 +851,10 @@ void tessAddContour( TESStesselator *tess, int size, const void* vertices,
 		* vertices in such an order that a CCW contour will add +1 to
 		* the winding number of the region inside the contour.
 		*/
-        e->winding = tess->reverseContours ? -1 : 1;
-        e->Sym->winding = tess->reverseContours ? 1 : -1;
+		e->winding = 1;
+		e->Sym->winding = -1;
 	}
 }
-
-void tessSetOption( TESStesselator *tess, int option, int value )
-{
-	switch(option)
-	{
-	case TESS_CONSTRAINED_DELAUNAY_TRIANGULATION:
-		tess->processCDT = value > 0 ? 1 : 0;
-		break;
-	case TESS_REVERSE_CONTOURS:
-		tess->reverseContours = value > 0 ? 1 : 0;
-		break;
-	}
-}
-
 
 int tessTesselate( TESStesselator *tess, int windingRule, int elementType,
 				  int polySize, int vertexSize, const TESSreal* normal )
@@ -1015,7 +876,7 @@ int tessTesselate( TESStesselator *tess, int windingRule, int elementType,
 	}
 
 	tess->vertexIndexCounter = 0;
-
+	
 	if (normal)
 	{
 		tess->normal[0] = normal[0];
@@ -1030,7 +891,7 @@ int tessTesselate( TESStesselator *tess, int windingRule, int elementType,
 	if (vertexSize > 3)
 		vertexSize = 3;
 
-	if (setjmp(tess->env) != 0) {
+	if (setjmp(tess->env) != 0) { 
 		/* come back here if out of memory */
 		return 0;
 	}
@@ -1064,9 +925,7 @@ int tessTesselate( TESStesselator *tess, int windingRule, int elementType,
 	if (elementType == TESS_BOUNDARY_CONTOURS) {
 		rc = tessMeshSetWindingNumber( mesh, 1, TRUE );
 	} else {
-		rc = tessMeshTessellateInterior( mesh );
-		if (rc != 0 && tess->processCDT != 0)
-			tessMeshRefineDelaunay( mesh, &tess->alloc );
+		rc = tessMeshTessellateInterior( mesh ); 
 	}
 	if (rc == 0) longjmp(tess->env,1);  /* could've used a label */