+++ docs/Changelog.md
@@ -2,11 +2,15 @@
 
 ## Next version
 
+- Android native libs for 16 KB page size [#1179](https://github.com/mapsforge/vtm/pull/1179)
 - Multi-map `MapFileTileSource.setPriority` [#1176](https://github.com/mapsforge/vtm/pull/1176)
 - Color filter theme resources [#1175](https://github.com/mapsforge/vtm/pull/1175)
   - `ThemeCallback.getBitmap`
+- Rename `MapDatabase.readLabels` to `readNamedItems` [#1181](https://github.com/mapsforge/vtm/pull/1181)
+  - Rename `MapDatabase.Selector.LABELS` to `NAMED`
 - Update JTS, OkHttp dependencies [#1172](https://github.com/mapsforge/vtm/pull/1172) [#1173](https://github.com/mapsforge/vtm/pull/1173)
 - Remove SLF4J dependency [#1167](https://github.com/mapsforge/vtm/pull/1167) [#1170](https://github.com/mapsforge/vtm/pull/1170)
+- Libtess2 v1.0.1 [#1178](https://github.com/mapsforge/vtm/pull/1178)
 - Disable `vtm-ios` [#1166](https://github.com/mapsforge/vtm/pull/1166)
 - Remove `vtm-theme-comparator` [#1169](https://github.com/mapsforge/vtm/pull/1169)
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.24.0)